Add preprocessor option '-x'
Some projects (like the strongSwan project) use macros to some magic
when it comes to arguemnt passing to functions. As a result a function
might not look like a regular function would. cscope does not recognize
these functions an cannot build appropriate reference files. The option
introduced allows for the macros to be removed by a 'real' function
allowing the proper use of cscope.
The patch is not ideal (it does a fork for each file in case it needs to
be preprocessd) and makes use of various of these nasty global variables
(which are already abundant across the project).
Any other ideas (as suggested), like modifying the lex/yacc files to
build a separate version for any project that needs preprocessing is
insanely stupid and lacks any understanding of software develepment (or
common sense). This needed to be said!

Signed-off-by: Thomas Egerer <hakke_007@gmx.de>